Commands now run simultaneous & clean exit possible
Added a new implementation of LongPollingBot (files under api/) that
starts every command as a new thread. To use that implementation changes
to DisplayBot (DisplayBot.java and *Parser.java) and some commands were necessary.
With that change I also added an onClosing method to the bot to be able
to shut the bot down cleanly and let the commands finish their work.
Additionally the main method now has a mechanism to stop the bot: every
 five seconds it is looking for a file that is called 'stop' in its working directory,
 when this file exists the program exits. The file can be created with a script called stopBot(.bat)
 that is added by gradle (build.gradle + stopBot*StartScript.txt).