applied new bot model to commands
Changed the structure of the Bot and the commands to use the new
thread for sending messages and the added api implementations (see a762f41 and 9b64985).

Also changed parsing of documents, because Telegram API has a strange handling of video and photo files,
e. g. some photos are treated as photos, some are treated as documents, the same applies for videos.